fixed bootstrap to take more samples
I thought the problem might be that there weren't enough entropy samples
per survey response to determine outliers, so I changed the number of
bootstrap iterations (previously 100) to 100000*some fraction (taken
from emery's C++ metric). However, due to the magnitude of the sampling
I can't get it to finish running on my computer, so I can't actually
confirm that was the problem.