First attempt at survey representation
Skeleton of the code:

Doesn't include all behavior yet (like responses, blocking, branching).
I'm not sure all the functions are actually necessary: the majority of
them could be removed and instance variables of the classes could be
directly referenced/manipulated by the user, but I thought the functions
might make it clearer/easier to the user how to create the survey (could
be wrong, i don't know). Haven't decided between repl or static creation
of survey (or both) yet.